Rework `ancestry_expected_proof_size` (#7)
* Adjust node proofs verification logic

* Compute optimal ancestry proof size

* Add comments

* Use BTreeSet/BTreeMap

* use entry API

* calculate proof size from flipped sub-mmr

Calculating the number of copath items required for the proof of the
`prev_peaks` living beneath some peak can be done
simultaneously more succinctly & efficiently as follows:
Each peak is a perfect merkle tree. Therefore if we have some `prev_peaks`
living beneath a given `peak`, then the number of co-path items to
prove said `prev_peaks` equals the number of peaks that would be formed
for a sub-mmr with `leaf_count = leaves(peak) - leaves(prev_peaks)`. This is akin to
flipping the construction of the sub-mmr to be right-to-left and
counting its peaks.

* add benchmarks for ancestry & node proofs

* add proof size logic from #6 for benching

* Revert "add proof size logic from #6 for benching"

This reverts commit 6cd3808c7d4738aeee4845979ba197756113b331.

* improve readability with slice comparison

Suggested-by: Serban Iorga <serban@parity.io>

* Revert "Compute optimal ancestry proof size"

This reverts commit 5e0fa6eaa7bdfcb48493099a773d87c3841265f8.

* Revert "Adjust node proofs verification logic"

This reverts commit dc1f5e89a7f229bcfb40c5a29e9f5cb6ed5de097.

* add tests: "Adjust node proofs verification logic"

Co-authored-by: Serban Iorga <serban@parity.io>

* add tests: "Compute optimal ancestry proof size"

Co-authored-by: Serban Iorga <serban@parity.io>

* add applicable comment from "Add comments"

Co-authored-by: Serban Iorga <serban@parity.io>

---------

Co-authored-by: Serban Iorga <serban@parity.io>